Have the war task be part of the pipeline [#3]
Instead of calling 'uber and 'jar, and writing outside the fileset, the
'immutant-war task should instead let the user construct the fileset
they need, and then add in any needed bits for immutant. Then the user
can pass that fileset to the 'war task to generate a war.

This is still a WIP, and may change somewhat.