Make the plugin a proper boot citizen
The plugin was a straight port of the leiningen one, so did things in a
non-booty way, like creating the uberjar and war itself, and writing
outside the fileset.

This fixes that, instead writing to the fileset instead of creating a
war, letting the standard war task assemble the war. It also changes the
test task to not build a war, but instead use a given war.

This also also renames immutant-war to gird, since it doesn't really
build a war anymore, just preps the fileset for doing so, and renames
immutant-test to test-in-container.